fix: controllers

+++ internal/application/dto/bid/response.go
@@ -5,8 +5,8 @@ import "time"
 type InstallationRequestResponse struct {
 	ID             uint      `json:"id"`
 	UserID         uint      `json:"userId"`
-	Area           float64   `json:"area"`
-	PowerRequested float64   `json:"powerRequested"`
+	Area           uint      `json:"area"`
+	PowerRequested uint      `json:"powerRequested"`
 	MaxCost        float64   `json:"maxCost"`
 	Deadline       time.Time `json:"deadline"`
 	BuildingType   string    `json:"buildingType"`
